% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_nomes_compostos.R
\name{obter_dic_nomes_proprios_compostos}
\alias{obter_dic_nomes_proprios_compostos}
\title{Carrega o dataset nomes_proprios_compostos 
Verifica se o dataset nomes_proprios_compostos existe em uma pasta de cache local.
Se inexistir, baixa-o de um release do GitHub e o salva no cache.}
\usage{
obter_dic_nomes_proprios_compostos()
}
\value{
Um data.table contendo os dados de nomes_proprios_compostos.
}
\description{
Carrega o dataset nomes_proprios_compostos 
Verifica se o dataset nomes_proprios_compostos existe em uma pasta de cache local.
Se inexistir, baixa-o de um release do GitHub e o salva no cache.
}
\keyword{internal}
