% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{remove_acentos_numeros_e_capitaliza}
\alias{remove_acentos_numeros_e_capitaliza}
\title{Pré-processamento Fonético: Remove acentos, números e capitaliza}
\usage{
remove_acentos_numeros_e_capitaliza(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres pré-processado.
}
\description{
Remove diacríticos, converte para maiúsculas e remove caracteres
que não sejam letras ou espaços.
}
\examples{
# nomesbr:::remove_acentos_numeros_e_capitaliza("João Silva 123!")
}
\keyword{internal}
