% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{remove_espacos_e_letras_duplicadas}
\alias{remove_espacos_e_letras_duplicadas}
\title{Remoção Fonética: Espaços e Letras Duplicadas}
\usage{
remove_espacos_e_letras_duplicadas(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres limpo.
}
\description{
Remove letras duplicadas (exceto RR, SS que podem ter sons distintos) e espaços extras.
}
\examples{
# nomesbr:::remove_espacos_e_letras_duplicadas(c("  CARRO OSSO  ", "JOAOZINHO"))
}
\keyword{internal}
