% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{remove_letras_silenciosas}
\alias{remove_letras_silenciosas}
\title{Remoção Fonética: Letras Silenciosas}
\usage{
remove_letras_silenciosas(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres com 'H's iniciais removidos.
}
\description{
Remove 'H' silencioso no início de cada palavra.
}
\examples{
# nomesbr:::remove_letras_silenciosas("Helena Horta")
}
\keyword{internal}
