% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simplificacao_nomes.R
\name{remove_PARTICULAS_AGNOMES}
\alias{remove_PARTICULAS_AGNOMES}
\title{Remove Partículas, Agnomes e algumas Patentes de Nomes}
\usage{
remove_PARTICULAS_AGNOMES(s)
}
\arguments{
\item{s}{Vetor de caracteres contendo nomes.}
}
\value{
Vetor de caracteres com nomes simplificados.
}
\description{
Remove Partículas, Agnomes e algumas Patentes de Nomes
}
\examples{
vct_nomes <- c("JOAO DA SILVA FILHO","CORONEL JACINTO")
remove_PARTICULAS_AGNOMES(vct_nomes)

}
