% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{remove_vogais_duplicadas}
\alias{remove_vogais_duplicadas}
\title{Remoção Fonética: Vogais Duplicadas}
\usage{
remove_vogais_duplicadas(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres com vogais duplicadas removidas.
}
\description{
Comprime sequências de vogais idênticas adjacentes.
}
\examples{
# nomesbr:::remove_vogais_duplicadas(c("AARAO", "REEBECA"))
}
\keyword{internal}
