% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{simplifica_consoantes_similares}
\alias{simplifica_consoantes_similares}
\title{Simplificação Fonética: Consoantes Similares}
\usage{
simplifica_consoantes_similares(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres com consoantes simplificadas.
}
\description{
Agrupa sons consonantais foneticamente similares.
}
\examples{
# nomesbr:::simplifica_consoantes_similares("CAÇADOR CELSO CARLOS GEOVANIA WALTER YARA ZEBRA")
}
\keyword{internal}
