% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{simplifica_digrafos_similares}
\alias{simplifica_digrafos_similares}
\title{Simplificação Fonética: Dígrafos Similares}
\usage{
simplifica_digrafos_similares(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres com dígrafos simplificados.
}
\description{
Transforma dígrafos comuns para simplificar a representação fonética.
}
\examples{
# nomesbr:::simplifica_digrafos_similares("FILHA MANHA CHICO SCHMIDT SCENA ESCOVA QUILO")
}
\keyword{internal}
