% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simplificacao_nomes.R
\name{simplifica_PARTICULAS_AGNOMES_PATENTES}
\alias{simplifica_PARTICULAS_AGNOMES_PATENTES}
\title{Cria coluna com agnomes, algumas patentes/cargos as remove, remove partículas}
\usage{
simplifica_PARTICULAS_AGNOMES_PATENTES(d, s = "nome_clean")
}
\arguments{
\item{d}{um objeto `data.table`}

\item{s}{string com nome da coluna de caracteres contendo nomes para 
simplificar. Por padrão, "nome_clean".}
}
\value{
data.table com novas colunas de nome simplificado e de marca  
agnomes_titulos
}
\description{
Cria coluna com agnomes, algumas patentes/cargos as remove, remove partículas
}
\examples{
dt_nomes <- data.table::data.table(nome = c("JOAO DA SILVA FILHO",
"CORONEL JACINTO"))
dt_nomes <- simplifica_PARTICULAS_AGNOMES_PATENTES(d=dt_nomes,s="nome")
print(dt_nomes)

}
