% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_metaphoneBR.R
\name{simplifica_sons_nasais_terminais}
\alias{simplifica_sons_nasais_terminais}
\title{Simplificação Fonética: Sons Nasais Terminais}
\usage{
simplifica_sons_nasais_terminais(texto)
}
\arguments{
\item{texto}{Um vetor de caracteres.}
}
\value{
Um vetor de caracteres com sons nasais terminais simplificados.
}
\description{
Unifica sons nasais no final das palavras.
}
\examples{
# nomesbr:::simplifica_sons_nasais_terminais(c("JOAQUIN", "JOAQUIM"))
}
\keyword{internal}
