% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomesbr-package.R
\docType{package}
\name{nomesbr-package}
\alias{nomesbr}
\alias{nomesbr-package}
\title{nomesbr: Limpa e Simplifica Nomes de Pessoas (Name Cleaner and Simplifier)}
\description{
Limpa e simplifica nomes de pessoas para auxiliar no pareamento de banco de dados na ausência de chaves únicas não ambíguas. Detecta e corrige erros tipográficos mais comuns, simplifica opcionalmente termos sujeitos eventualmente a omissão em cadastros, e simplifica foneticamente suas palavras, aplicando variação própria do algoritmo metaphoneBR. (Cleans and simplifies person names to assist in database matching when unambiguous unique keys are unavailable. Detects and corrects common typos, optionally simplifies terms prone to omission in records, and applies phonetic simplification using a custom variation of the metaphoneBR algorithm.) Mation (2025) \doi{10.6082/uchicago.15104}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeadata-lab/nomesbr}
  \item \url{https://ipeadata-lab.github.io/nomesbr/}
  \item Report bugs at \url{https://github.com/ipeadata-lab/nomesbr/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Borges \email{rodrigoesborges@gmail.com} (\href{https://orcid.org/0000-0003-2076-1424}{ORCID})

Authors:
\itemize{
  \item Lucas Mation (\href{https://orcid.org/0000-0002-7461-932X}{ORCID})
}

Other contributors:
\itemize{
  \item Ipea - Institue for Applied Economic Research [copyright holder, funder]
}

}
\keyword{internal}
