% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simplificacao_nomes.R
\name{simplifica_PARTICULAS_AGNOMES_PATENTES}
\alias{simplifica_PARTICULAS_AGNOMES_PATENTES}
\title{Cria coluna com agnomes, algumas patentes/cargos as remove, remove partículas}
\usage{
simplifica_PARTICULAS_AGNOMES_PATENTES(d, s = "nome_clean")
}
\arguments{
\item{d}{um objeto `data.table`}

\item{s}{Vetor de caracteres contendo nomes para simplificar. Padrão nome_clean.}
}
\value{
Vetor de caracteres com nomes simplificados.
}
\description{
Cria coluna com agnomes, algumas patentes/cargos as remove, remove partículas
}
